// condoplanner-app/src/apiClient/models/UpsertUserCondominiumInput.ts
/* tslint:disable */
/* eslint-disable */
/**
 * API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { ETipoUsuario } from './ETipoUsuario';
import {
    ETipoUsuarioFromJSON,
    ETipoUsuarioFromJSONTyped,
    ETipoUsuarioToJSON,
    ETipoUsuarioToJSONTyped,
} from './ETipoUsuario';

/**
 * 
 * @export
 * @interface UpsertUserCondominiumInput
 */
export interface UpsertUserCondominiumInput {
    /**
     * 
     * @type {number}
     * @memberof UpsertUserCondominiumInput
     */
    condominiumId?: number;
    /**
     * 
     * @type {number}
     * @memberof UpsertUserCondominiumInput
     */
    userId?: number;
    /**
     * 
     * @type {ETipoUsuario}
     * @memberof UpsertUserCondominiumInput
     */
    userType?: ETipoUsuario;
    /**
     * 
     * @type {boolean}
     * @memberof UpsertUserCondominiumInput
     */
    active?: boolean;
    /**
     * 
     * @type {Date}
     * @memberof UpsertUserCondominiumInput
     */
    startDate?: Date | null;
}



/**
 * Check if a given object implements the UpsertUserCondominiumInput interface.
 */
export function instanceOfUpsertUserCondominiumInput(value: object): value is UpsertUserCondominiumInput {
    return true;
}

export function UpsertUserCondominiumInputFromJSON(json: any): UpsertUserCondominiumInput {
    return UpsertUserCondominiumInputFromJSONTyped(json, false);
}

export function UpsertUserCondominiumInputFromJSONTyped(json: any, ignoreDiscriminator: boolean): UpsertUserCondominiumInput {
    if (json == null) {
        return json;
    }
    return {
        
        'condominiumId': json['condominiumId'] == null ? undefined : json['condominiumId'],
        'userId': json['userId'] == null ? undefined : json['userId'],
        'userType': json['userType'] == null ? undefined : ETipoUsuarioFromJSON(json['userType']),
        'active': json['active'] == null ? undefined : json['active'],
        'startDate': json['startDate'] == null ? undefined : (new Date(json['startDate'])),
    };
}

export function UpsertUserCondominiumInputToJSON(json: any): UpsertUserCondominiumInput {
    return UpsertUserCondominiumInputToJSONTyped(json, false);
}

export function UpsertUserCondominiumInputToJSONTyped(value?: UpsertUserCondominiumInput | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'condominiumId': value['condominiumId'],
        'userId': value['userId'],
        'userType': ETipoUsuarioToJSON(value['userType']),
        'active': value['active'],
        'startDate': value['startDate'] == null ? value['startDate'] : value['startDate'].toISOString(),
    };
}